LRQA-32058 Move getting date information into DateUtil
