+++ modules/test/poshi-runner/src/main/java/com/liferay/poshi/runner/util/DateUtil.java
@@ -16,6 +16,7 @@ package com.liferay.poshi.runner.util;
 
 import java.text.SimpleDateFormat;
 
+import java.util.Calendar;
 import java.util.Date;
 import java.util.Locale;
 
@@ -24,19 +25,94 @@ import java.util.Locale;
  */
 public class DateUtil {
 
-	public static String getCurrentDate(String pattern, Locale locale) {
-		return getDate(new Date(), pattern, locale);
+	@Override
+	public String getCurrentDay() {
+		Calendar calendar = Calendar.getInstance();
+
+		return StringUtil.valueOf(calendar.get(Calendar.DATE));
 	}
 
-	public static String getDate(Date date, String pattern, Locale locale) {
-		SimpleDateFormat simpleDateFormat = new SimpleDateFormat(
-			pattern, locale);
+	@Override
+	public String getCurrentDayName() {
+		Calendar calendar = Calendar.getInstance();
 
-		return simpleDateFormat.format(date);
+		return StringUtil.valueOf(
+			calendar.getDisplayName(
+				Calendar.DAY_OF_WEEK, Calendar.LONG, Locale.US));
+	}
+
+	@Override
+	public String getCurrentHour() {
+		Calendar calendar = Calendar.getInstance();
+
+		return StringUtil.valueOf(calendar.get(Calendar.HOUR_OF_DAY));
+	}
+
+	@Override
+	public String getCurrentMonth() {
+		Calendar calendar = Calendar.getInstance();
+
+		return StringUtil.valueOf(calendar.get(Calendar.MONTH) + 1);
+	}
+
+	@Override
+	public String getCurrentMonthName() {
+		Calendar calendar = Calendar.getInstance();
+
+		return String.valueOf(calendar.getDisplayName(
+			Calendar.MONTH, Calendar.LONG, Locale.US));
+	}
+
+	@Override
+	public String getCurrentYear() {
+		Calendar calendar = Calendar.getInstance();
+
+		return StringUtil.valueOf(calendar.get(Calendar.YEAR));
+	}
+
+	@Override
+	public String getDay(String days) {
+		Calendar calendar = Calendar.getInstance();
+		calendar.add(Calendar.DATE, Integer.valueOf(days));
+
+		return StringUtil.valueOf(calendar.get(Calendar.DATE));
+	}
+
+	@Override
+	public String getDayName(String days) {
+		Calendar calendar = Calendar.getInstance();
+		calendar.add(Calendar.DATE, Integer.valueOf(days));
+
+		return StringUtil.valueOf(
+			calendar.getDisplayName(
+				Calendar.DAY_OF_WEEK, Calendar.LONG, Locale.US));
+	}
+
+	@Override
+	public String getMonth(String days) {
+		Calendar calendar = Calendar.getInstance();
+		calendar.add(Calendar.DATE, Integer.valueOf(days));
+
+		return StringUtil.valueOf(calendar.get(Calendar.MONTH) + 1);
+	}
+
+	@Override
+	public String getMonthName(String days) {
+		Calendar calendar = Calendar.getInstance();
+		calendar.add(Calendar.DATE, Integer.valueOf(days));
+
+		return String.valueOf(calendar.getDisplayName(
+			Calendar.MONTH, Calendar.LONG, Locale.US));
 	}
 
 	public static String getNanoseconds() {
 		return String.valueOf(System.nanoTime());
 	}
 
+	@Override
+	public String getYear(String days) {
+		Calendar calendar = Calendar.getInstance();
+		calendar.add(Calendar.DATE, Integer.valueOf(days));
+
+		return StringUtil.valueOf(calendar.get(Calendar.YEAR));
 }
